mini commit #15

+++ environment/hanabi_learning_environment/agents/train.py
@@ -105,7 +105,7 @@ def run_training(
             if done:
                 agent.memory.push(
                     torch.FloatTensor(observation["vectorized"]),
-                    torch.FloatTensor(action_number),
+                    torch.LongTensor(action_number),
                     None,
                     reward,
                 )
@@ -113,7 +113,7 @@ def run_training(
             else:
                 agent.memory.push(
                     torch.FloatTensor(observation["vectorized"]),
-                    torch.FloatTensor(action_number),
+                    torch.LongTensor(action_number),
                     torch.FloatTensor(new_obs["vectorized"]),
                     reward,
                 )
